refactor(admin): split pipeline board modules
Extract board model constants and draggable board subcomponents so pipeline-board.tsx owns drag sensors, optimistic status updates, and composition only.

Verified with pnpm exec prettier --check apps/admin/src/app/content/pipeline-board.tsx apps/admin/src/app/content/pipeline-board-components.tsx apps/admin/src/app/content/pipeline-board-model.ts; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/content/pipeline-board.tsx apps/admin/src/app/content/pipeline-board-components.tsx apps/admin/src/app/content/pipeline-board-model.ts; pnpm validate; pnpm dlx knip --no-exit-code.
