+++ apps/admin/src/app/content/pipeline-board.tsx
@@ -1,7 +1,6 @@
 "use client";
 
-import { useState, useCallback, useRef, useEffect } from "react";
-import { useRouter } from "next/navigation";
+import { useState, useCallback, useRef } from "react";
 import {
   DndContext,
   DragOverlay,
@@ -10,256 +9,24 @@ import {
   KeyboardSensor,
   useSensor,
   useSensors,
-  defaultDropAnimationSideEffects,
 } from "@dnd-kit/core";
 import type { DragStartEvent, DragEndEvent } from "@dnd-kit/core";
-import {
-  SortableContext,
-  verticalListSortingStrategy,
-  useSortable,
-} from "@dnd-kit/sortable";
-import { useDroppable } from "@dnd-kit/core";
-import { CSS } from "@dnd-kit/utilities";
 import { updateContentStatus } from "../actions";
-import { STATUS_COLORS, SERIES_COLORS, PLATFORM_ABBREV } from "@/lib/constants";
 import type { ContentStatus } from "@anipotts/types";
+import {
+  DraggableCard,
+  DroppableColumn,
+  ProgressBar,
+  Toast,
+} from "./pipeline-board-components";
+import {
+  COLUMNS,
+  dropAnimation,
+  type BoardItem,
+  type PipelineBoardProps,
+} from "./pipeline-board-model";
 
-export interface BoardItem {
-  id: string;
-  title: string;
-  status: string;
-  series_type: string | null;
-  atom_count: number;
-  platforms_posted: string[];
-  updated_at: string;
-}
-
-interface PipelineBoardProps {
-  items: BoardItem[];
-}
-
-const COLUMNS: ContentStatus[] = [
-  "idea",
-  "draft",
-  "ready",
-  "atomized",
-  "published",
-];
-
-const COLUMN_HEADER_COLORS: Record<string, string> = {
-  idea: "bg-zinc-700",
-  draft: "bg-yellow-500/20",
-  ready: "bg-blue-500/20",
-  atomized: "bg-purple-500/20",
-  published: "bg-green-500/20",
-};
-
-// ── Progress Bar ──
-
-function ProgressBar({ items }: { items: BoardItem[] }) {
-  const counts: Record<string, number> = {};
-  for (const col of COLUMNS) {
-    counts[col] = items.filter((i) => i.status === col).length;
-  }
-  const total = items.length || 1;
-
-  const segments = COLUMNS.map((col) => ({
-    status: col,
-    count: counts[col] || 0,
-    pct: ((counts[col] || 0) / total) * 100,
-  }));
-
-  return (
-    <div className="px-1 pb-3">
-      <div className="flex h-1.5 rounded-full overflow-hidden bg-zinc-800/60">
-        {segments.map((seg) =>
-          seg.count > 0 ? (
-            <div
-              key={seg.status}
-              className={`${COLUMN_HEADER_COLORS[seg.status]} transition-all duration-300`}
-              style={{ width: `${seg.pct}%` }}
-            />
-          ) : null,
-        )}
-      </div>
-      <p className="text-[10px] text-zinc-500 mt-1.5">
-        {segments
-          .filter((s) => s.count > 0)
-          .map((s) => `${s.count} ${s.status}`)
-          .join(" \u00b7 ")}
-      </p>
-    </div>
-  );
-}
-
-// ── Draggable Card ──
-
-function DraggableCard({
-  item,
-  overlay,
-}: {
-  item: BoardItem;
-  overlay?: boolean;
-}) {
-  const router = useRouter();
-  const wasDragged = useRef(false);
-
-  const {
-    attributes,
-    listeners,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({
-    id: item.id,
-    data: { status: item.status },
-  });
-
-  const style = overlay
-    ? undefined
-    : {
-        transform: CSS.Transform.toString(transform),
-        transition,
-      };
-
-  const date = new Date(item.updated_at).toLocaleDateString("en-US", {
-    month: "short",
-    day: "numeric",
-  });
-
-  function handlePointerDown() {
-    wasDragged.current = false;
-  }
-
-  function handlePointerMove() {
-    wasDragged.current = true;
-  }
-
-  function handleClick() {
-    if (!wasDragged.current) {
-      router.push(`/content/${item.id}`);
-    }
-  }
-
-  const baseClasses =
-    "rounded-lg border border-zinc-800/60 bg-zinc-900/50 p-3 cursor-grab active:cursor-grabbing";
-  const draggingClasses = isDragging ? "opacity-40 scale-[0.98]" : "";
-  const overlayClasses = overlay
-    ? "ring-2 ring-violet-500/50 shadow-lg shadow-violet-500/10"
-    : "";
-
-  return (
-    <div
-      ref={overlay ? undefined : setNodeRef}
-      style={style}
-      className={`${baseClasses} ${draggingClasses} ${overlayClasses}`}
-      onPointerDown={handlePointerDown}
-      onPointerMove={handlePointerMove}
-      onClick={handleClick}
-      {...(overlay ? {} : attributes)}
-      {...(overlay ? {} : listeners)}
-    >
-      <p className="text-[13px] text-zinc-200 leading-snug line-clamp-2">
-        {item.title}
-      </p>
-
-      <div className="flex items-center gap-1 mt-2 flex-wrap">
-        {item.atom_count > 0 && (
-          <span className="admin-badge bg-zinc-800 text-zinc-400">
-            {item.atom_count}a
-          </span>
-        )}
-        {item.series_type && (
-          <span
-            className={`admin-badge ${SERIES_COLORS[item.series_type as keyof typeof SERIES_COLORS] || ""}`}
-          >
-            {item.series_type}
-          </span>
-        )}
-        {item.platforms_posted.length > 0 &&
-          item.platforms_posted.map((p) => (
-            <span
-              key={p}
-              className="admin-badge bg-green-500/10 text-green-400"
-            >
-              {PLATFORM_ABBREV[p] || p}
-            </span>
-          ))}
-      </div>
-
-      <p className="text-[10px] text-zinc-600 mt-2 tabular-nums">{date}</p>
-    </div>
-  );
-}
-
-// ── Droppable Column ──
-
-function DroppableColumn({
-  status,
-  items,
-}: {
-  status: ContentStatus;
-  items: BoardItem[];
-}) {
-  const { setNodeRef } = useDroppable({ id: status });
-  const itemIds = items.map((i) => i.id);
-
-  return (
-    <div className="w-72 shrink-0 flex flex-col">
-      <div className="flex items-center gap-2 px-2 pb-2">
-        <span
-          className={`admin-badge ${STATUS_COLORS[status] || STATUS_COLORS.draft}`}
-        >
-          {status}
-        </span>
-        <span className="text-[10px] text-zinc-600 tabular-nums">
-          {items.length}
-        </span>
-      </div>
-
-      <div
-        ref={setNodeRef}
-        className="flex-1 min-h-[300px] max-h-[500px] overflow-y-auto admin-scroll rounded-lg border border-zinc-800/40 bg-zinc-900/20 p-2 space-y-2"
-      >
-        <SortableContext items={itemIds} strategy={verticalListSortingStrategy}>
-          {items.length === 0 && (
-            <p className="text-[11px] text-zinc-600 text-center py-8">
-              No items
-            </p>
-          )}
-          {items.map((item) => (
-            <DraggableCard key={item.id} item={item} />
-          ))}
-        </SortableContext>
-      </div>
-    </div>
-  );
-}
-
-// ── Toast ──
-
-function Toast({ message, onDone }: { message: string; onDone: () => void }) {
-  useEffect(() => {
-    const t = setTimeout(onDone, 3000);
-    return () => clearTimeout(t);
-  }, [onDone]);
-
-  return (
-    <div className="fixed bottom-4 right-4 z-50 rounded-lg border border-red-500/30 bg-red-950/90 px-4 py-2 text-[12px] text-red-300 shadow-lg">
-      {message}
-    </div>
-  );
-}
-
-// ── Board ──
-
-const dropAnimation = {
-  duration: 150,
-  sideEffects: defaultDropAnimationSideEffects({
-    styles: { active: { opacity: "0.4" } },
-  }),
-};
+export type { BoardItem } from "./pipeline-board-model";
 
 export default function PipelineBoard({
   items: initialItems,
@@ -278,11 +45,11 @@ export default function PipelineBoard({
   );
 
   const activeItem = activeId
-    ? (items.find((i) => i.id === activeId) ?? null)
+    ? (items.find((item) => item.id === activeId) ?? null)
     : null;
 
   const columnItems = useCallback(
-    (status: string) => items.filter((i) => i.status === status),
+    (status: string) => items.filter((item) => item.status === status),
     [items],
   );
 
@@ -299,28 +66,24 @@ export default function PipelineBoard({
     const overId = String(over.id);
     let targetStatus: string;
 
-    // If dropped on a column, overId is the status string
     if (COLUMNS.includes(overId as ContentStatus)) {
       targetStatus = overId;
     } else {
-      // Dropped on a card, find that card's status
-      const overItem = items.find((i) => i.id === overId);
+      const overItem = items.find((item) => item.id === overId);
       if (!overItem) return;
       targetStatus = overItem.status;
     }
 
-    const draggedItem = items.find((i) => i.id === String(active.id));
+    const draggedItem = items.find((item) => item.id === String(active.id));
     if (!draggedItem || draggedItem.status === targetStatus) return;
 
-    // Optimistic update
     const prevItems = items;
     setItems((prev) =>
-      prev.map((i) =>
-        i.id === draggedItem.id ? { ...i, status: targetStatus } : i,
+      prev.map((item) =>
+        item.id === draggedItem.id ? { ...item, status: targetStatus } : item,
       ),
     );
 
-    // Server call
     isUpdating.current = true;
     try {
       const result = await updateContentStatus(
